Delete Collections.java - all methods now replaced by native Java
