--- modules/common/src/main/java/org/opencastproject/util/data/Collections.java
@@ -1,37 +1,0 @@
-/*
- * Licensed to The Apereo Foundation under one or more contributor license
- * agreements. See the NOTICE file distributed with this work for additional
- * information regarding copyright ownership.
- *
- *
- * The Apereo Foundation licenses this file to you under the Educational
- * Community License, Version 2.0 (the "License"); you may not use this file
- * except in compliance with the License. You may obtain a copy of the License
- * at:
- *
- *   http://opensource.org/licenses/ecl2.txt
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
- * License for the specific language governing permissions and limitations under
- * the License.
- *
- */
-
-package org.opencastproject.util.data;
-
-/**
- * This class provides functions to ease and secure the handling of collections by supporting a type safe -- at least to
- * the extent Java's type system allows -- immutable and more functional style.
- *
- * Note that all functions do <i>not</i> mutate input collections unless otherwise stated.
- *
- * @deprecated By now, all of this is natively available in Java.
- */
-@Deprecated
-public final class Collections {
-  private Collections() {
-  }
-
-}
